requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
bcc==0.1.10
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.2
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==2.2.4
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
stripe==12.2.0
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
